+++ src/RcsbGraphQL/QueryTest.ts
@@ -4,7 +4,7 @@ import {CoreEntry} from "./Types/Yosemite/GqlTypes";
 import fs from "fs";
 import {AlignmentResponse, SequenceReference} from "./Types/Borrego/GqlTypes";
 
-const yosemiteClient: GraphQLRequest = new GraphQLRequest("yosemite" , {fetch: fetch as any});
+const yosemiteClient: GraphQLRequest = new GraphQLRequest("data-api" , {fetch: fetch as any});
 const yosemiteQuery = fs.readFileSync(__dirname+"/Queries/Yosemite/QueryMultipleEntriesProperties.graphql", "utf-8");
 const yoemiteRequest = async ()=>{
     const response = await yosemiteClient.request<{entryIds:string[]},{entries:CoreEntry[]}>(
@@ -16,7 +16,7 @@ const yoemiteRequest = async ()=>{
     console.log(response);
 }
 
-const borregoClient: GraphQLRequest = new GraphQLRequest("borrego" , {fetch: fetch as any});
+const borregoClient: GraphQLRequest = new GraphQLRequest("1d-coordinates" , {fetch: fetch as any});
 const borregoQuery = fs.readFileSync(__dirname+"/Queries/Borrego/QueryAlignments.graphql", "utf-8");
 const borregoRequest = async ()=>{
     const response = await borregoClient.request<{queryId: string; from: SequenceReference, to:SequenceReference},{alignment:AlignmentResponse}>(