+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 [Semantic Versioning](https://semver.org/)
 
+## [4.0.0] - 2022-03-23
+### Breaking change
+- Replaced graphql client `@apollo/client` for `graphql-request`
+- `GraphQLRequest` constructor parameters changed
+  - `api: "data-api" | "1d-coordinates" | string`
+    - `"data-api"` requests to RCSB data-api (https://data.rcsb.org).
+    - `"1d-coordinates"` requests to RCSB 1d-coordinate service (https://1d-coordinates.rcsb.org
+    - `string` encoding a URL 
+  - `config` additional configuration for `GraphQLClient` ('graphql-request' npm package https://www.npmjs.com/package/graphql-request) 
+  
 ## [3.0.1] - 2022-03-17
 ### Improvement
 - Library adds caching and data compression at browser side through `localStorage`